backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
scikit-learn==1.5.2
xgboost==2.1.1
numpy==1.26.4
pandas==2.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
aiosqlite==0.20.0
psycopg2-binary==2.9.9
pytest==8.3.5
APScheduler==3.10.4
